Improve derivative registration.
Once the derivative's FunctionDecl is created it needs to be registered properly
in the AST. Before this patch `registerDerivative` ignored many checks which
make sure the produced code is valid C++. This patch enforces more rigorous
checks for some declaration kinds such as standalone functions and some cases
of class functions. Lambdas will be handled in subsequent patches.

The challenge is to produce valid C++ code when differentiating classes.
Consider:
```cpp
class A1 {
  double f(double x) { return x * x; }
  double f_grad(double x, double *d_x); // forward declaration of the gradient.
};
```

In such cases clad will now produce a correct out-of-line definition of `f_grad`:
`double A1::f_grad(double x, double *d_x) { ... }`.

In cases where the gradient function is not declared as part of the class
signature (external libraries such as STL), we continue to handle them as
before. That will be dealt with in a subsequent patches.

The drawback of the patch is that we continue to produce copy pasted code in
`Derive`, `DerivePullback` and `DerivePushforward`. This will be cleaned up next.